+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/game/GameSolvedListener.kt
@@ -1,5 +1,5 @@
 package org.piepmeyer.gauguin.game
 
 fun interface GameSolvedListener {
-    fun puzzleSolved(troughReveal: Boolean)
+    fun puzzleSolved()
 }