+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/undo/UndoManager.kt
@@ -2,41 +2,17 @@ package org.piepmeyer.gauguin.undo
 
 import org.piepmeyer.gauguin.grid.GridCell
 
-class UndoManager(private val listener: UndoListener) {
-    private val undoList = mutableListOf<UndoState>()
+interface UndoManager {
+    fun addListener(listener: UndoListener)
 
-    fun clear() {
-        undoList.clear()
-    }
+    fun clear()
 
     fun saveUndo(
         cell: GridCell,
         batch: Boolean,
-    ) {
-        val undoState =
-            UndoState(
-                cell,
-                cell.userValue,
-                cell.possibles,
-                batch,
-            )
-        undoList.add(undoState)
-        listener.undoStateChanged(true)
-    }
+    )
 
-    fun restoreUndo() {
-        if (undoList.isNotEmpty()) {
-            val undoState = undoList.removeLast()
-            val cell = undoState.cell
-            cell.setUserValueIntern(undoState.userValue)
-            cell.possibles = undoState.possibles
-            cell.isLastModified = true
-            if (undoState.isBatch) {
-                restoreUndo()
-            }
-        }
-        if (undoList.isEmpty()) {
-            listener.undoStateChanged(false)
-        }
-    }
+    fun restoreUndo()
+
+    fun undoPossible(): Boolean
 }