+++ gauguin-core/src/main/kotlin/org/piepmeyer/gauguin/undo/UndoStep.kt
@@ -2,7 +2,7 @@ package org.piepmeyer.gauguin.undo
 
 import org.piepmeyer.gauguin.grid.GridCell
 
-data class UndoState(
+data class UndoStep(
     val cell: GridCell,
     val userValue: Int,
     val possibles: Set<Int>,