+++ gauguin-app/src/main/kotlin/org/piepmeyer/gauguin/ui/main/MainActivity.kt
@@ -209,7 +209,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun configureActivity() {
-        activityUtils.configureNightMode()
+        activityUtils.configureTheme(this)
         activityUtils.configureKeepScreenOn(this)
         activityUtils.configureFullscreen(this)