+++ gauguin-app/src/main/kotlin/org/piepmeyer/gauguin/ui/statistics/StatisticsActivity.kt
@@ -43,6 +43,7 @@ class StatisticsActivity : AppCompatActivity() {
             resetStatisticsDialog()
         }
 
+        activityUtils.configureTheme(this)
         activityUtils.configureFullscreen(this)
 
         scatterPlotDiagramFragment = StatisticsScatterPlotDiagramFragment()