+++ CHANGELOG.md
@@ -11,10 +11,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [0.33.0] - 2024-12-21
+
+### Changed
+
 * Moved UI for theme switching from preferences screen to the main menu. Just open the burger menu
   and find the button straight on top.
-
-### Deprecated
 
 ### Removed
 
@@ -23,8 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 * Fix theme switching if either the old or the new theme is the dynamic theme from Material You.
-
-### Security
 
 ## [0.32.0] - 2024-11-10